My saddest commit ever. Purging ts files and converting to javascript.
