+++ index.html
@@ -18,6 +18,7 @@
   <script src="js/dcr_parser.js"></script>  
   <script src="js/GUI.js"></script>  
   <script src="js/communication.js"></script>
+  <script src="js/Simulation.js"></script>
 
   <table id="task-table"></table>  
   <p id="accepting"></p>
