+++ js/Simulation.js
@@ -1,51 +1,54 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const parser = require("./dcr_parser");
 class Simulation {
+
     constructor(input) {
-        this.isRunning = false;
         this.graph = parser.parse(input);
         this.users = [];
         this.id = 1;
+        this.isRunning = false;
+        this.startTime = undefined
+        this.stopTime = undefined
+        this.users = []
         this.log = [];
     }
+
     executeEvent(event) {
-        this.graph.execute(event);
+        this.graph.execute(event)
         this.log.push(event);
     }
+
     startSimulation() {
         this.isRunning = true;
         this.startTime = new Date();
     }
+
     stopSimulation() {
         this.isRunning = false;
         this.stopTime = new Date();
     }
+
     addUsers(user) {
         this.users.push(user);
     }
+
     saveLog() {
         console.log(this.log);
     }
+
     hello() {
         if (this.isRunning) {
             console.log("Hello, i AM running :)");
-        }
-        else {
+        } else {
             console.log("Hello, i am NOT running :(");
         }
     }
 }
-var sim = new Simulation((`A(0,0,0)      
-B(0,1,1)        
-A -->* B
-B *--> A
-C -->% A
-D -->+ A    
-D -->* B
-A --><> (B, D)`));
-sim.hello();
-sim.startSimulation();
-sim.hello();
-sim.stopSimulation();
-sim.hello();
+
+
+
+class User {
+    constructor(id, name) {
+        this.id = id;
+        this.name = name;
+        this.roles = []
+    }
+}
