--- js/User.js
@@ -1,10 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.User = void 0;
-class User {
-    constructor(id, name) {
-        this.id = id;
-        this.name = name;
-    }
-}
-exports.User = User;
